# swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "MY TEST API",
    "description": "TEST API with Swagger-autogen",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3030/"
    }
  ],
...
  "paths": {
...
    "/auth/sign-in": {
      "post": {
        "tags": [
          "auth"
        ],
        "description": "유저 회원가입 엔드포인트",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "유저 정보",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/AddUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "회원가입에 성공했습니다!",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "example": "any"
                  },
                  "password": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/sign-in": {
      "post": {
        "tags": [
          "auth"
        ],
        "description": "유저 회원가입 엔드포인트",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "유저 정보",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/AddUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "회원가입에 성공했습니다!",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "example": "any"
                  },
                  "password": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Parents": {
        "type": "object",
        "properties": {
          "father": {
            "type": "string",
            "example": "Simon Doe"
          },
          "mother": {
            "type": "string",
            "example": "Marie Doe"
          }
        },
        "xml": {
          "name": "Parents"
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Jhon Doe"
          },
          "age": {
            "type": "number",
            "example": 29
          },
          "parents": {
            "xml": {
              "name": "parents"
            },
            "$ref": "#/components/schemas/Parents"
          },
          "diplomas": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "school": {
                  "type": "string",
                  "example": "XYZ University"
                },
                "year": {
                  "type": "number",
                  "example": 2020
                },
                "completed": {
                  "type": "boolean",
                  "example": true
                },
                "internship": {
                  "type": "object",
                  "properties": {
                    "hours": {
                      "type": "number",
                      "example": 290
                    },
                    "location": {
                      "type": "string",
                      "example": "XYZ Company"
                    }
                  }
                }
              }
            }
          }
        },
        "xml": {
          "name": "User"
        }
      },
      "AddUser": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "haechan"
          },
          "password": {
            "type": "string",
            "example": "1234"
          }
        },
        "required": [
          "username",
          "password"
        ],
        "xml": {
          "name": "AddUser"
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "in": "header",
        "bearerFormat": "JWT"
      }
    }
  }
}